+++ app/src/main/java/com/wxxtfxrmx/spritzreader/domain/entity/Book.kt
@@ -6,7 +6,7 @@ data class Book(
 	val path: String,
 	val name: String,
 	val cover: String? = null,
-	val progress: Int = 0,
+	val progress: Float = 0f,
 	val description: String = "",
 	val lastPage: Int = 0,
 	val wordOnPage: Int = 0
